refactor repository
