+++ litestar/contrib/sqlalchemy/repository/_util.py
@@ -1,0 +1,31 @@
+from __future__ import annotations
+
+from contextlib import contextmanager
+from typing import Any
+
+from sqlalchemy.exc import IntegrityError, SQLAlchemyError
+
+from litestar.contrib.repository import ConflictError, RepositoryError
+
+
+@contextmanager
+def wrap_sqlalchemy_exception() -> Any:
+    """Do something within context to raise a `RepositoryError` chained
+    from an original `SQLAlchemyError`.
+
+        >>> try:
+        ...     with wrap_sqlalchemy_exception():
+        ...         raise SQLAlchemyError("Original Exception")
+        ... except RepositoryError as exc:
+        ...     print(f"caught repository exception from {type(exc.__context__)}")
+        ...
+        caught repository exception from <class 'sqlalchemy.exc.SQLAlchemyError'>
+    """
+    try:
+        yield
+    except IntegrityError as exc:
+        raise ConflictError from exc
+    except SQLAlchemyError as exc:
+        raise RepositoryError(f"An exception occurred: {exc}") from exc
+    except AttributeError as exc:
+        raise RepositoryError from exc
