+++ pyproject.toml
@@ -420,3 +420,12 @@ known-first-party = ["litestar", "tests", "examples"]
 "litestar/params.py" = ["N802"]
 "test_apps/**/*.*" = ["D", "TRY", "EM", "S", "PTH"]
 "tools/**/*.*" = ["D", "ARG", "EM", "TRY", "G", "FBT"]
+
+
+[tool.unasyncd]
+files = {"litestar/contrib/sqlalchemy/repository/async_repository.py" = "litestar/contrib/sqlalchemy/repository/sync_repository.py"}
+
+[tool.unasyncd.per_file_add_replacements."litestar/contrib/sqlalchemy/repository/async_repository.py"]
+SQLAlchemyAsyncRepository = "SQLAlchemySyncRepository"
+"sqlalchemy.ext.asyncio.AsyncSession" = "sqlalchemy.orm.Session"
+"litestar.contrib.repository.AbstractAsyncRepository" = "litestar.contrib.repository.AbstractSyncRepository"
